+++ src/lib/sfio/Stdio_b/ftell.c
@@ -5,16 +5,16 @@
 */
 
 #if __STD_C
-long ftell(reg FILE *fp)
+long ftell(reg FILE* fp)
 #else
 long ftell(fp)
-reg FILE	*fp;
+reg FILE*	fp;
 #endif
 {
-	reg Sfio_t	*sp;
+	reg Sfio_t*	sp;
 
 	if(!(sp = _sfstream(fp)))
 		return -1L;
 	_stdclrerr(fp,sp);
-	return sftell(sp);
+	return (long)sftell(sp);
 }
